Fix dmsgget tests by removing obsolete default value check
Tests were failing on Appveyor's slow builds such as Apple because the
Mock discovery API checked if the timeout in entries was set to zero.
The dmsg client does not update automatically anymore but due to a
default value check we were setting the timeout in entry to default
value leading to the entry being cleaned and the download client being
unable to connect to the download server.