Add dynamic dmsg-servers
This commit removes the static dmsg-server in the code and adds a way to add dmsg-servers dynamically via PostEntry() in directClient. PostEntry() can also be used to add a client entry to the direct client where all the servers in the direct client will also be added to the delegatedServers filed of the client entry. AllServers is added in the dmsg-discovery api to get all the servers connected to the discovery. The NewDirectClient is moved from StartDmsg to before dmsghttp.ListenAndServe so that the servers and entries can be added dynamically. A new util file is created in dmsghttp with GetServers and GetServers to get servers form the dmsg-discovery and update them periodically.