Close stream with defer
This commit adds a way to close the stream of dmsg via a defer statement with a go routine in it inside the fun RoundTrip().

The goroutine checks if the response body is closed every one second and if it is closed then we clsoe the stream if not then we keep checking till it's closed or if context is closed.