Minor fixes and changes
This commit contains some minor linting fixes.

It also removes a test log and renames the func test used in RoundTrip() to closeStream() and adds a warn log in faliure to close stream and adds a comment detailing how we check if the response body is closed.

The todo by evanlinjin is also removed as we won't be reusing the streams as we are now closing them after every use.