Fix closeStream
This commit fixes closeStream() in dmsghttp. The error check is changed from using err strings to using errors.IS to check the error http.ErrBodyReadAfterClose. Also a new error is added to the check, io.EOF.

Now we check if there is an error and if the error is equal to either http.ErrBodyReadAfterClose or io.EOF.