Fix outdated timestamp error
This commit fixes the outdated timestamp error during the shutdown of the visor.

As the client entries no longer update themselves in the dmsg-discovery from time to time, the timestamp of the entry is never updated and so it gives the outdated timestamp error in delEntry() if it is called 1 minute after the clent has been registered.

We fix the issue by simply not checking the timestamp if the entry is client in delEntry().