Fix reconnection edgecase
This commit fixes the reconnection edgecase where if the dmsg-server that we are trying to connect to is the last one and the method ce.EnsureSession returns an error then the reconnect  gets stuck on the select statement waiting for either <-ce.done or <-ce.errCh and only moves when the done chan is closed.

In order to fix that we send the error of the method ce.EnsureSession if the dmsg-server that we are trying to connect to is the last one.