Fix noise read block
This commit fixes the read block caused in noise InitiatorHandshake when a dmsg server is closed improperly.

Instead of using the func InitiatorHandshake we use the the method Handshake of ReadWriter which already has a timeout logic.

The method Buffered is added to ReadWriter as it was needed in initClient and server to check the number of bytes that can be read from the buffer rawInput.