Fix appveyor windows build
This commit fixes the windows build on app veyor by increasing the ReadTimeout, WriteTimeout and ReadHeaderTimeout to 3 seconds.

This windows build used to fail on appveyor but the same check worked on local windows machine likey due to the appveyor having less resources hence slowing the run and the server timing out because of the 1 second timeouts.