Add customCommands to urls commands
This commit adds a way to pass custom commands to the pty as a paramater to the Handler().

This commit removes the previously hardcoded custom command 'update' that was used for updating skywire. The reason for this change was for this we had to import skywire for the commands which created a cycle where dmsg imported skywire and skywire imported dmsg which is not good.

This also gives us flexibility to add more custom commands if required.