feat: Add DialServerForIP method to dmsg client
This commit adds a new method `DialServerForIP` to the `Client` struct in the `pkg/dmsg/client.go` file. This method dials to dmsg servers to retrieve the public IP address of the client. It iterates through a list of server entries, attempts to dial each server, and returns the first public IP address it receives.

The purpose of this change is to provide a way for the client to obtain its public IP address from dmsg servers.