Feat/ip from server (#272)
* feat: Add DialServerForIP method to dmsg client

This commit adds a new method `DialServerForIP` to the `Client` struct in the `pkg/dmsg/client.go` file. This method dials to dmsg servers to retrieve the public IP address of the client. It iterates through a list of server entries, attempts to dial each server, and returns the first public IP address it receives.

The purpose of this change is to provide a way for the client to obtain its public IP address from dmsg servers.

* feat: Add new commandline tool dmsgip

* chore: fix linting

* feat: Add support for dmsg server public keys in DialServerForIP method

This commit modifies the `DialServerForIP` method in the `pkg/dmsg/client.go` file to accept a slice of dmsg server public keys as an argument. If the `servers` argument is nil, the method retrieves the server entries using the `discoverServers` function and populates the `servers` slice with the static public keys from the entries. Then, it iterates through the `servers` slice and attempts to dial each server to retrieve the public IP address of the client.

The purpose of this change is to allow the `DialServerForIP` method to support custom dmsg server public keys, providing more flexibility in obtaining the client's public IP address from dmsg servers.

* feat: Add error handling for non-public IP address in DialServerForIP method

* refactor: Improve error handling and connection logic in DialServerForIP method

This commit refactors the `DialServerForIP` method in the `pkg/dmsg/client.go` file to improve error handling and connection logic. It introduces two separate loops to handle delegated servers and attempts to connect to each server individually. Additionally, it properly closes if the session is created just for the IP after dialing the server.

The purpose of this change is to enhance the reliability and stability of the `DialServerForIP` method when retrieving the public IP address from dmsg servers.

* refactor: Improve error handling and add MinSessions to startDmsg

* feat: Add dmsgip commandline tool to dmsg asa a subcommand

* refactor: Rename DialServerForIP to LookupIP in dmsg client

This commit renames the `DialServerForIP` method to `LookupIP` in the `pkg/dmsg/client.go` file. The functionality remains the same, but the new name better reflects the purpose of the method, which is to lookup the public IP address of the client from dmsg servers.

The purpose of this change is to improve the clarity and consistency of the method name, making it more intuitive for developers working with the dmsg client.

* chore: Move code around

* refactor: Improve error handling in LookupIP method

* test: add unit tests for LookupIP

* refactor: improve error handling in dmsgip cmd

* test: update IP lookup logic in stream_test.go

This commit updates the IP lookup logic in the `stream_test.go` file. It introduces conditional checks based on the operating system to handle different IP address formats. On Windows, the IP address is expected to be "127.0.0.1", while on other operating systems, it is expected to be "::1".

The purpose of this change is to ensure that the IP lookup tests pass correctly on different operating systems, improving the reliability and consistency of the test suite.